# Row accumulator support update Scalar values