@yahoNanJing 2023-04-14
<meta charset="utf-8"><b style="font-weight:normal;" id="docs-internal-guid-41cd3865-7fff-3961-439b-1a53bcdd67c7"><div dir="ltr" style="margin-left:0pt;" align="left">

TPCH 10g | 2023-03-20(d77ccc25979e3a5e7d71ca86a6317fce50a4bff7) | 2023-04-12(​​2c573105c3d3e4044b4ca74e4faa59d31e57f17b)https://github.com/apache/arrow-datafusion/pull/5866https://github.com/apache/arrow-datafusion/pull/5837 | 2023-04-14(​​bd705fe6bb6f018b71dc1f25409e45ec061fd6df)https://github.com/apache/arrow-datafusion/pull/5973 | 2023-04-14(5bfc1410d677b3da2b9a4adaa3f032562a14a113)https://github.com/apache/arrow-datafusion/pull/6003
-- | -- | -- | -- | --
q1 | 23665.52 | 20753.61 | 20891.52 | 21408.54
q2 | 5589.88 | 3617.40 | 3464.09 | 3464.95
q3 | 9777.48 | 5961.30 | 5785.48 | 5732.81
q4 | 5188.19 | 3226.53 | 3132.99 | 4091.51
q5 | 15309.18 | 12133.36 | 10995.68 | 14361.51
q6 | 6293.92 | 4114.41 | 4504.54 | 3529.13
q7 | 28215.12 | 25501.06 | 23664.61 | 24111.81
q8 | 11563.27 | 8585.95 | 10085.88 | 7257.50
q9 | 19719.88 | 16851.29 | 16703.78 | 16446.74
q10 | 14240.17 | 9811.16 | 9618.89 | 10092.34
q11 | 5774.94 | 4358.94 | 4326.38 | 4629.20
q12 | 7027.51 | 4771.94 | 4659.69 | 4712.21
q13 | 17182.12 | 15855.66 | 15594.57 | 16362.75
q14 | 7567.25 | 4924.66 | 4679.48 | 4458.82
q15 | 17523.50 | 11911.70 | 11484.73 | 8626.74
q16 | 4064.01 | 3719.74 | 3707.31 | 3667.75
q17 | 135964.77 | 117756.03 | 107820.88 | 83682.60
q18 | 89023.33 | 60923.75 | 65235.46 | 66134.85
q19 | 9244.44 | 8504.44 | 11795.41 | 8504.76
q20 | 25746.50 | 15681.54 | 15023.04 | 12232.71
q21 | 34265.73 | 33962.40 | 34681.59 | 34787.75
q22 | 8026.30 | 8000.52 | 8334.53 | 8487.91
Total | 495383.13 | 400927.39 | 396190.53 | 366784.89

</div></b>