@mingmwang 2023-04-14
Q5

Before this PR

Running benchmarks with the following options: DataFusionBenchmarkOpt { query: Some(5), debug: false, iterations: 3, partitions: 1, batch_size: 8192, path: "./parquet_data10", file_format: "parquet", mem_table: false, output_path: None, disable_statistics: true, enable_scheduler: false }
Query 5 iteration 0 took 4093.0 ms and returned 5 rows
Query 5 iteration 1 took 4031.0 ms and returned 5 rows
Query 5 iteration 2 took 4008.9 ms and returned 5 rows
Query 5 avg time: 4044.29 ms


After this PR

Running benchmarks with the following options: DataFusionBenchmarkOpt { query: Some(5), debug: false, iterations: 3, partitions: 1, batch_size: 8192, path: "./parquet_data10", file_format: "parquet", mem_table: false, output_path: None, disable_statistics: true, enable_scheduler: false }
Query 5 iteration 0 took 4042.3 ms and returned 5 rows
Query 5 iteration 1 took 3978.7 ms and returned 5 rows
Query 5 iteration 2 took 3973.1 ms and returned 5 rows
Query 5 avg time: 3998.02 ms

Almost no difference.